requirements.txt
pyautogui>=0.9.54
pyperclip>=1.8.2
customtkinter>=5.2.0
python-docx>=1.1.0
PyMuPDF>=1.24.0
